refactor: Update data type for goods order info from Integer to Long
- Modified GoodsInfoTypeHandler to parse Long instead of Integer for goods order IDs
- Updated BaseGoodsOrdersDTO, BaseGoodsOrdersVO, and GoodsOrders to use Map<Long, Double>
- Aligned type handling across multiple classes to support larger ID ranges
- Improved type consistency in goods order information processing

+++ src/main/java/com/sakurapuare/boatmanagement/pojo/dto/base/BaseGoodsOrdersDTO.java
@@ -13,6 +13,6 @@ public class BaseGoodsOrdersDTO extends BaseOrderDTO {
 
 
     @ApiModelProperty("订单信息：id:数量,id:数量,id:数量")
-    private Map<Integer, Double> orderInfo;
+    private Map<Long, Double> orderInfo;
 }
